feat : 쿠폰 목록조회(관리자) / 장바구니 생성 구현

+++ src/main/java/com/ecommerce/chatdemo/domain/wishlist/dto/WishListResponse.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -16,7 +15,7 @@ public class WishListResponse {
     private Long memberId;
     private Long productId;
     private String productName;
-    private BigDecimal productPrice;
+    private Long productPrice;
     private LocalDateTime createdAt;
 
     public WishListResponse(WishList wishList) {
